+++ app/ui/employee-details.tsx
@@ -7,14 +7,14 @@ import {
   FormItem,
   FormLabel,
   FormMessage,
-} from '@/components/ui/form';
+} from '@/app/ui/form';
 import {
   HoverCard,
   HoverCardContent,
   HoverCardTrigger,
-} from '@/components/ui/hover-card';
-import { Button } from '@/components/ui/button';
-import { Input } from '@/components/ui/input';
+} from '@/app/ui/hover-card';
+import { Button } from '@/app/ui/button';
+import { Input } from '@/app/ui/input';
 
 import { formSchema } from '../schemas/formSchemas';
