+++ app/ui/work-details.tsx
@@ -7,15 +7,15 @@ import {
   FormItem,
   FormLabel,
   FormMessage,
-} from '@/components/ui/form';
+} from '@/app/ui/form';
 import {
   HoverCard,
   HoverCardContent,
   HoverCardTrigger,
-} from '@/components/ui/hover-card';
-import { ToggleGroup, ToggleGroupItem } from '@/components/ui/toggle-group';
-import { Button } from '@/components/ui/button';
-import { Input } from '@/components/ui/input';
+} from '@/app/ui/hover-card';
+import { ToggleGroup, ToggleGroupItem } from '@/app/ui/toggle-group';
+import { Button } from '@/app/ui/button';
+import { Input } from '@/app/ui/input';
 
 import { Day, formSchema } from '../schemas/formSchemas';
